requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
pymongo==4.0.1
slack_sdk==3.11.2
python-dotenv==0.19.2
Flask==2.3.3
